% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMCA_gui.R
\name{scMCA_vis}
\alias{scMCA_vis}
\title{Visualize scMCA result in a web browser.}
\usage{
scMCA_vis(mca_result)
}
\arguments{
\item{mca_result}{an object generate by scMCA funtion}
}
\description{
Runs interactive \code{shiny} session of \code{scMCA}
}
